requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.0.9
django-bootstrap4==2.2.0
django-countries==7.1
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
